requirements.txt
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
databases==0.8.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.108.0
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.10
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.32.0.post1
strawberry-graphql==0.217.1
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
